webapp/python/requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.26
SQLAlchemy==1.4.22
PyJWT[crypto]==2.1.0